+++ tests/integration/studies_blueprint/test_study_copy.py
@@ -14,8 +14,8 @@ from pathlib import Path
 
 from starlette.testclient import TestClient
 
-from tests.integration.test_helpers.outputs import create_minimal_output_zip_from_name
 from tests.integration.utils import wait_task_completion
+from tests.test_helpers.outputs import create_minimal_output_zip_from_name
 
 
 def test_copy_with_editor_preservation(client: TestClient, admin_access_token: str) -> None:
