+++ tests/output/storage/file/test_file_output_storage.py
@@ -13,7 +13,6 @@ import os
 import shutil
 import tempfile
 import zipfile
-from datetime import datetime, timezone
 from pathlib import Path
 from unittest.mock import Mock
 
@@ -45,6 +44,7 @@ from antarest.study.storage.rawstudy.model.filesystem.config.files import build
 from antarest.study.storage.rawstudy.model.filesystem.config.model import Mode
 from antarest.study.storage.rawstudy.model.filesystem.factory import FileStudy
 from antarest.study.storage.rawstudy.model.filesystem.root.filestudytree import FileStudyTree
+from tests.test_helpers.dates import utc_to_local
 
 
 @pytest.fixture
@@ -391,12 +391,6 @@ def _extract_output_to_dir(study_zip: Path, output_path: str, target_dir: Path) 
         (tmp_dir / "STA-mini" / "output" / "20201014-1422eco-hello").rename(target_dir)
 
 
-def _utc_to_local_date(date_str: str) -> str:
-    """Converts a UTC date string such as "20201014-1222" to a local date string such as "20201014-1422"."""
-    utc_date = datetime.strptime(date_str, "%Y%m%d-%H%M").replace(tzinfo=timezone.utc)
-    return utc_date.astimezone().strftime("%Y%m%d-%H%M")
-
-
 def test_import_output_directory(
     file_output_storage: InStudyFileOutputStorage, tmp_path: Path, sta_mini_zip_path: Path
 ) -> None:
@@ -413,7 +407,7 @@ def test_import_output_directory(
     file_output_storage.import_output("my-study", output_dir)
 
     # Cannot hard code the expected formatted date because it depends on the locale
-    expected_date = _utc_to_local_date("20201014-1222")
+    expected_date = utc_to_local("20201014-1222")
 
     expected_output_dir = study_dir / "output" / f"{expected_date}eco-hello"
     assert expected_output_dir.exists()
@@ -458,7 +452,7 @@ def test_import_output_zip_should_import_it_as_archived(
     output_id = file_output_storage.import_output("my-study", zip_path)
 
     # Cannot hard code the expected formatted date because it depends on the locale
-    expected_date = _utc_to_local_date("20201014-1222")
+    expected_date = utc_to_local("20201014-1222")
 
     assert output_id == f"{expected_date}eco-hello"
     assert file_output_storage.list_outputs("my-study") == [
@@ -512,7 +506,7 @@ def test_import_output_archive_stream(
         file_output_storage.import_output("my-study", f)
 
     # Cannot hard code the expected formatted date because it depends on the locale
-    expected_date = _utc_to_local_date("20201014-1222")
+    expected_date = utc_to_local("20201014-1222")
 
     assert file_output_storage.list_outputs("my-study") == [
         OutputMetadata(id=f"{expected_date}eco-hello", in_study=True, archived=False)
