+++ tests/storage/business/test_export.py
@@ -36,7 +36,7 @@ from antarest.study.storage.variantstudy.model.command_context import CommandCon
 from tests.conftest import empty_study_fixture
 from tests.db_statement_recorder import DBStatementRecorder
 from tests.helpers import create_raw_study, dirhash, with_db_context
-from tests.integration.test_helpers.outputs import create_minimal_output_dir_from_name
+from tests.test_helpers.outputs import create_minimal_output_dir_from_name
 
 
 def test_export_flat_export_all_files_except_output(
